Fixed some tests. Many to go. Existing project roadwaynumber assingment / splitting not yet handled.
Small Changes to changetable handling.
Some fixes after running tests.
