+++ digiroad2-viite/src/main/scala/fi/liikennevirasto/viite/process/strategy/TrackCalculatorStrategy.scala
@@ -152,19 +152,6 @@ trait TrackCalculatorStrategy {
       case _ =>
           (averageOfAddressMValues(rightLink.startAddrMValue, leftLink.startAddrMValue, reversed), averageOfAddressMValues(rightLink.endAddrMValue, leftLink.endAddrMValue, reversed))
     }
-
-//    (leftLink.status, rightLink.status) match {
-//      case (LinkStatus.Transfer, LinkStatus.Transfer) | (LinkStatus.UnChanged, LinkStatus.UnChanged) =>
-//        (averageOfAddressMValues(rightLink.startAddrMValue, leftLink.startAddrMValue, reversed), averageOfAddressMValues(rightLink.endAddrMValue, leftLink.endAddrMValue, reversed))
-//      case (LinkStatus.UnChanged, _) | (LinkStatus.Transfer, _) =>
-//        (leftLink.startAddrMValue, leftLink.endAddrMValue)
-//      case (_, LinkStatus.UnChanged) | (_, LinkStatus.Transfer) =>
-//        (rightLink.startAddrMValue, rightLink.endAddrMValue)
-//      case _ =>
-//        userCalibrationPoint.map(c => (c.addressMValue, c.addressMValue)).getOrElse(
-//          (averageOfAddressMValues(rightLink.startAddrMValue, leftLink.startAddrMValue, reversed), averageOfAddressMValues(rightLink.endAddrMValue, leftLink.endAddrMValue, reversed))
-//        )
-//    }
   }
 
   protected def setLastEndAddrMValue(projectLinks: Seq[ProjectLink], endAddressMValue: Long): Seq[ProjectLink] = {
